Add custom types to collectd server
When trying to introduce a new plugin to collect rabbitmq stats,
we discovered that
- the collectd server exists
- and it needs the same typesdb as the client,
   in order to decode the stats it receives

We introduce a config file which lists the types files required and move the
typesdb file for the rabbitmq plugin so they are available to the server as
well as all clients.
Clients which do not install the collectd-rabbitmq plugin have no use for
the additional types but suffer no harm, and now there is one true home
for any future custom types files.

@matmoore and @benhyland