Upgrade puppetlabs/rabbitmq module
To allow us to delete rabbitmq queue bindings from puppet, we need
https://github.com/puppetlabs/puppetlabs-rabbitmq/pull/340
which is not in our current version.

Fix the module release version in the Puppetfile to avoid pulling in tip,
which would introduce a dependency on the unreleased puppetlabs-stdlib tip.

See https://trello.com/c/eP2CdnrC/590-fix-rabbitmq-queue-binding