Sync Whitehall attachments to S3
This adds functionality to also backup Whitehall attachments to S3 by
adding `s3cmd` to the loop in which attachments are copied over to the
asset-slaves if they are detected as being free from viruses.

It is uploaded to S3 in the same directory structure as to the
asset-slaves to help with consistency and make it easier to restore the
directory structure if required. The nightly sync which also happens (to
ensure that the master and slaves are in sync is also updated.

The process which creates the `process-uploaded-attachments.sh` and
`copy-attachments.sh` scripts are moved to asset-master only due to the way
that they had to incorporate s3_bucket credentials.