moves govuk_crawler_worker rabbitmq to 'mirrorer'
The crawler has been known to crash the production rabbitmq.
It has beed decided to move the rabbitmq process away from
the rabbitmq cluster and to create a local rabbitmq instance
on the mirrorer app server. This will effectively make the
govuk_crawler_worker app self contained and should lighten
the load on the rabbitmq cluster