Add monitoring for archiving
When we use WAL-E for backups we switch on the WAL continuous archiving
option
(http://www.postgresql.org/docs/9.3/static/continuous-archiving.html).
Everytime a WAL segment is archived ends up in S3, and the archive
timeout option is set to 60 seconds, which means we will push something
to S3 every 60 seconds.

This adds a wrapper around the command which includes a passive check
set to a threshold of 15 minutes. If a WAL segment hasn't been archived
for over 15 minutes then we'll be alerted.