Add dashboard to hold graphs for publishing api queues
We introduce a graph for panopticon only, non-erb-templated, as a first pass.
We will tidy and add extra graphs soon.

This graph shows all the stats that might be reported by the
govuk_message_queue_consumer gem.

It ensure that the height of the graph is always representing actual stat values,
by using summarize and consolidateBy(max) graphite functions.
As a result, it is possible for the area under the graph to not accurately represent
the running total correctly - but this is not expected by the users.
The tradeoff is necessary due to the aggregation and bucketing done by statsd
and graphite, and by the possibility that there might be fewer pixels than buckets.

@jackscotti and @benhyland