Use new deployment repo for post-deploy repo backup job
All of the shared deployment code from alphagov-deployment, including the
sync_to_ghe.sh script which this job uses, is already in the new repo:
https://github.com/alphagov/govuk-app-deployment/blob/master/sync_to_ghe.sh

The Deploy_App job uses the new repo already so we should update this job too
for consistency. This change brings us nearer to getting all deployment code
out of alphagov-deployment.