Add script to resync environment
This adds a defined type to create a wrapper script to allow syncing from an
S3 bucket from another environment to the current server. It uses WAL-E
backups which first pulls a base backup from the latest available
backup, adds in the required options for recovery and starts the
PostgreSQL server which will start up the recovery process.

I've left it simply as a script for now to keep it open about how we
wish to schedule the job.