Add example configuration
This is example configuration for adding the environment sync script to
transition_postgresql_master in Integration. It adds all the required
secrets to be able to pull a database stored elsewhere (i.e. a
Production bucket) and an enable flag so we only enable this in the
environment we want (as it won't ever need to be enabled in Production,
although the script might be useful to live there if we ever want to do
a quick rollback).