adds ES index snapshots to S3
This commit adds the ES plugin to AWS which allows us to snapshot
individual indices to an S3 bucket. S3 credentials can be passed
a variety of ways, but adding them to elasticsearch.yml provides
the simplest solution. Passing the creds as environment variables
requires changes to ES user and the ES init.d script and overrides
to the vendor module, and only works if these requirements are met.