Add logic to enable env sync
If an environment is syncing attachments from another environment, it
shouldn't also backup (considering the size and number of attachments
we're dealing with). This adds the logic to only enable backups on
environments that are not syncing, and simply adds the tool to allow
environment syncs. As in line with other work of this nature, just add
the tool to allow an environment sync rather than any other scheduling.

Also update the class documentation.