Set `$GOPATH` in the dev VM environment
This is part of the work underway to upgrade metadata-api to the latest
go version.

Currently the dev VM has no `$GOPATH` which requires the apps try and
build one within their own directory. This seems increasingly fragile
with the most recent >1.6 versions of go.

This commit sets the `$GOPATH` to `/var/govuk/gopath` which will require
our go apps to be located at
`/var/govuk/src/github.com/alphagov/<reponame>` at which point all will
be well in the world.