Bunch of bug fixes
I wasn't able to fix this up to the initial commit but can do if people
would prefer. I think it's good to be separate as it shows the bugs I
ran into when testing the script.

One of the main changes is to pass the GPG_PASSPHRASE in as an
environmental variable.