Add deployment job for Lambda apps
This commit adds a Jenkins job for Lambda apps. The idea behind it is to
grab the source from a new repository specifically for holding lambda
app info, push the content to an S3 bucket specifically created for
Lambda apps, and trigger a build of a Terraform job which is designed to
deploy the Lambda function in Amazon.