Update Nginx config to support specialist publisher rewrite
We plan on releasing specialist publisher one document_type at a
time. In order to facilitate this we need the capability of
routing specific paths to the 'rebuild' application whilst serving
the rest from the existing v1 application.

This updates the loadbalancer to direct cma-cases and
sp-rebuild/assets to the specialist-publisher-rebuild app.
It also sets the Host HTTP header to specialist-publisher for the
specialist-publisher-rebuild app.