Move attachment sync logic to s_asset_base
The asset master in Production was set up to do a full sync of
attachments to an S3 bucket. Unfortunately this caused the machine to
run out of memory and OOM killer to kill the process before it had
completed checking all the files on the source machine (completed about
40% before it ran OOM). We made the decision to move this process to an
asset slave instead where we can upgrade memory, perform maintenance and
will have less of an impact overall than the asset master.

However, because there will now be two machines sharing AWS credentials
it made sense to move the logic to s_asset_base, set enabling parameters
and specifically pick nodes to run the required function on. By default
we should have nothing that copies to S3 until we specifically enable it
in the node specific hieradata. Once enabled it should pick up the same
credentials so we don't have to repeat them in our secrets store.