Specifically enable asset sync functions
This commit serves to specifically enable certain functions across
different nodes. As a rule we only want Production to copy anything to
S3, and we want Staging and Integration to be able to sync their files
from the S3 bucket.

It also takes the bucket name out of the password store because
otherwise the credentials would be created across *all* asset machines
(because they all inherit s_asset_base) and we only want credentials on
the asset masters and one asset slave in Production.