WAL-E env sync bugfixes
In testing we found a few bugs which meant that this didn't work off the
bat.

One addition is to disable and then re-enable and re-run Puppet: because
we're restoring from a Production DB the passwords get updated, so
running Puppet ensures the passwords are correct. It will also ensure
that the gpg.conf file is updated and will remove the added passphrase.