Upgrade Licensify app servers to Oracle Java 8
This is part of the work that the new Licensify team are doing.

It will let them test Java 8 in integration.

The setup is purposefully more simple than with Java 7. We're not
caching the Java download anywhere. This may prove to be flaky and
we might have to revisit it.