Split out graphs for the different Sidekiq worker types in Rummager. This is nicer for comparing the consumer activity with sidekiq activity, since the consumer only ever uses the AmendWorker.
Graph the accepted/rejected counts of the rummager queue consumer.
This helps us understand discrepancies in the consumer activity vs the AmendWorker activity.

Avoid `sumSeries()`. This function is unnecessary since we only need to
sum into buckets, not sum multiple `seriesList`s. Note that `summarize()`
is already capable of summing multiple `series`, and a `series` is not
the same as a `seriesList`. `sumSeries()` has the unfortunate side effect
of averaging `series` where the sample rates are different, which in our
case is always. It doesn't make sense to show that sideqick processed a
non-integer fraction of a message in some bucket.

See https://github.com/alphagov/rummager/pull/655

@benhyland and @davidslv