Only require 1 elasticsearch master node on the dev vm
In 1cfda1339ca819fdeb80f8c027173e565c6e38dd we changed the default for
`minimum_master_nodes` to be 2.  This makes sense everywhere except on the
dev vm where we still want it to be 1.  If it remains as 2 then the dev
vm cluster is unhealthy and reports:

   {"error":"MasterNotDiscoveredException[waited for [30s]]","status":503}

for `curl -XGET 'localhost:9200/_cluster/health?pretty'`.  This in turn
means we cannot run replication scripts.  Setting it to 1 makes things
work again.