Clear integration redis after data sync
This commit runs `flushall` on redis-1 and redis-2 in the integration
environment following the completion of the data sync job.

Redis is predominantly used for Sidekiq queues. When running large
republishings on integration it is possible that there will be residual
jobs left in Redis when integration is switched off overnight. These
jobs will run when the environment comes back on the following morning
but will be running over different data than they were created against
possibly causing inconsistencies.

Both Redis hosts are included as they run independently and aren't
clustered.