Ensure passphrase in file is cleaned up
This was removed in 15c36348 under the assumption that Puppet will
remove the passphrase file because backups are enabled on the server
using [this
code](https://github.com/alphagov/govuk-puppet/blob/master/modules/govuk_postgresql/manifests/wal_e/backup.pp#L132-L138).

However, we may not want to enable backups on the server. If we add this
line back in *after* Puppet has run then at least we can be assured we
don't have a passphrase hanging around in a file (even though it's in an
env var file elsewhere).