I thought this might be helpful for others who have seen errors while running tests against puppet for code they have not touched.
` 1) govuk::node::s_cache by default configures varnish to strip cookies
     Failure/Error: is_expected.to contain_file('/etc/varnish/default.vcl').
     Puppet::Error:
       Invalid parameter preserve_fqdn on Class[Rsyslog] at
       /Users/USER/govuk/govuk-puppet/modules/govuk/manifests/node/s_base.pp:35
       on node cache-1.router.somethingsomething
`

I found that if i've had a live branch which I may have rebased for some time and in that time updated
my master branch. I get spec errors on code that I have not touched.  This is due to either missing or
outdated modules managed by librarian.  This does not occur on Jenkins or Travis because a new envirinment
is spawned each time a test is run.