Lowercase URLs if they are all capitals
Currently we serve 404s to users that have inadvertently entered a URL
with CAPSLOCK on. This commit downcases URLs that contain no lowercase
letters but ignores urls where the case is mixed.

This should reduce 404s received by users.

The location block regex in this commit matches `gov.uk/ / / /` and
causes a redirect loop for that 'url'. As it is invalid this shouldn't
be an issue.