Add notify_passive_check wrapper script
Local Links Manager requires several rake tasks to be run regularly.
This adds a more generic wrapper for communicating with Icinga on
success/failure.

We did this because Puppet knows the IP address of the machine the task
is running on and this means we do not need a script for each individual
task.

Publisher has a script for one task and we did not want to do the same
for Local Links Manager as this would have resulted in several scripts.