Remove jenkins jobs for local links manager imports
To keep everything consistent we'll be running these via cron jobs
controlled by the app, like we've just done for the link checker task.
We keep the icinga passive checks but move them to the app manifest
alongside the link checker one.  All of these are only created on
production (this controlled by the local_links_manager_passive_checks
flag).