Ensure mongo service name is correct
In later versions of MongoDB (3.x), and in the package we're installing
from specifically, the upstart service name has changed to 'mongod' from
'mongodb'. This causes a few problems in Puppet for dependencies, so as
a plaster before we upgrade fully to mongo 3 just specify the service
name.

This has been reworked to set the parameter names using a conditional if
the version is 3.x and above. It defaults to the older pre-3.x values as
this was set as the default and was not specifically overridden in our
code base.