Fix potential race condition in restore process
There is a potential race condition where the restore process might
start to download a backup which is in the process of being taken, which
means an incomplete backup will be downloaded and the restore will fail.

This introduces the process of pushing the backup into a temporary
filename, and then when that has fully completed, move the current
latest backup to an archive directory, and rename the temporary as the
current latest.