Mongodb 3.2 config file
Mongodb's config file format has changed since version 2.6 and will fail to start with without
the relevant configuration file.
The parameters are mostly the same between versions so we are using puppet
'versioncmp' method as the conditonal argument to apply the right file for
the version.
We added a 2nd template for mongodb version 3.2