Change dfid-transition app_type from `rack` to `procfile`
This means it won't start unicorns serving port 3124, so leaving the way
open for the "proper" channel of sidekiq-monitoring. Now we've added
this app on 3124 to sidekiq-monitoring's Procfile, sidekiq-monitoring is
attempting to rackup this app and causing a port collision, breaking
monitoring for other apps.

Let's be nicer than that.