Check existence of lock file before reboot
What
Unattended reboots when invoked, interrupt backups. We do not want to reboot the host while
a backup is in progress.

How
When a backup is initiated we create a file under /etc/unattended-reboot/no-reboot.
The file should only exists for the duration of the job. While the file exists,
the machine does not reboot.