Ensure Go apps are run from the correct path
Because we build Go apps when running them (rather than running a
pre-built binary), the source needs to be in the correct location
(within a workspace at $GOPATH) to build correctly.

This commit checks for the existence of the relevant directory in
the "wrong" place and moves it to the correct location if necessary.
It also creates a symlink from the new location to the old one and
notifies the user that this has happened.