Show a friendlier message if a path doesn't exist
Currently trying to bowl an app which isn't cloned results in a nasty
`No such file or directory` error, followed by the build command still
trying to run.  This change ensures that we first check that the
relevant directory exists before trying to move in to it, and provides
a nicer error message if it doesn't.