Show the reason Puppet is disabled in Icinga
This commit changes the check of the Puppet agent to also check whether
there's a lockfile on disk which is preventing Puppet from running.

Unfortunately this means that we need to make a fairly large change
to `govuk_puppet` which allows all exit codes to let the script continue
rather than just 0 or 2.