Extract db name correctly from new connect command
We expected the connect command in the postgres dump to be:

    \connect a-database-name

(with optional "s around the database name). This connect string has
changed recently (we think as part of this security patch[1]) to be:

    \connect -reuse-previous=on "dbname='a-database-name'"

Our script for extracting the name of the db would fail to extrat the
correct name and so the imports would fail.  This commit updates the sed
to extract the name from either of these forms as both seem to be present
in our dump files.

[1]: https://www.postgresql.org/about/news/1688/