Make Nginx proxy with HTTP/1.1 for Licensing
This commit adds a feature flag so that Nginx on the app servers can
proxy to the Licensing app on localhost using HTTP/1.1.

Nginx docs:

> By default, version 1.0 is used. Version 1.1 is recommended for use
> with keepalive connections and NTLM authentication.

http://nginx.org/en/docs/http/ngx_http_proxy_module.html#proxy_http_version

Licensing has been upgraded to a new version of Scala/Play which is picky
about which HTTP version it needs (because it uses chunked responses
for file downloads).