Add double-pipe at end of command
There was a problem where a file didn't exist for archiving. This caused
a script to bomb out and cause problems later down the line as it didn't
complete.

If there isn't an older backup to archive that's OK, so add a
double-pipe on the end of the command to ensure the script doesn't just
bomb out (in all other instances it's a good thing it does).

This also caused a problem where older backups weren't moved away and
where they should be files, they were turned into directories, so we had
incorrectly named backups (i.e.
`s3://bucket/latest/base.xbcrypt/base.xbcrypt/base.xbcrypt/base.xbcrypt`).