Change logic regarding previous backups
The base backup script shouldn't care if a backup has already been taken
because it creates a fresh backup in a fresh location, so this logic is
redundant (with the newer fallout logic for the archiving command).

The incremental backup should care if a base backup has ever been taken,
but this is likely to only ever be apparent on the first run of the
backup because nothing else removes the `/var/lib/mysql/xtrabackup_date`
file. Remove most of the logic, and wrap the function so it only ever
runs if a base backup has run full-stop.