Mongodb s3 Restore 8
prefixing 's3cmd' sub-command with 'envdir'.

While verifying the retore script an error occured stating that there was no '.s3cfg' file.
When invoking s3cmd it performs a lookup for aws auth tokens by checking for cmd-line arguments, env-variables and it's own config '.s3cfg'.
This occured while executing a sub-command that was not prefixed with the envdir command which is how we pass env-varibles.

Removed 'var/' sub-directory as this does not get created outside my testing environment.

This has been tested on a host running mongodb in integration.