Tweak Puppet run
In 73b49aa205b7226900ab90510cc2a6616edef865 I tried to get Icinga to
show the reason Puppet is disabled, but I had to remove the check
for exit code 2 on the Puppet run.

Puppet runs can have all kind of exit codes (especially when run with
`--detailed-exitcodes`) so we should probably let every exit code
through and catch problems as part of the script that checks the
run.

I've added an additional check that the run is recent enough.