MongoDB S3 Backup Tweaks
What
During verification on Staging, we noticed that tar was outputting "permission denied" on `/var/lib/s3backup/lost+found`.
This is because `/var/lib/s3backup` is now a mount point on the root of a separate volume.
This relates to https://github.com/alphagov/govuk-puppet/pull/4689

How
Exclude the `lost+found` directory by adding the `--exclude` flag to the `tar` command.