Use api-elasticsearch for calculators-frontend in integration
This is step one towards us getting rid of the 0.90 cluster.

The only deployed app that actually uses elasticsearch directly on
calculators-frontend is licencefinder and it can create it's index from
nothing so we can easily point the servers at a new cluster and create
the data again.

Tariff is still listed as an app for these machines, and it uses
elasticsearch, but it's since been deployed to the PaaS so we can ignore it.