Add backup and restore scripts to all environments
This commit reworks the hieradata to add backups to all s_mysql_backup
machines in all environments, and the restore script to all
s_mysql_master machines in all environments.

The logic to create the backups and scripts already existed and is gated
by the existance of the s3_bucket_name variable. The restore script uses
exactly the same credentials as the backup scripts within the same
environment, so it can just pinch the defined hieradata that we used for
backups.

This also adds some class documentation.