Amend logic of triggering the build
This commit fixes up a few bits that were copied across from the
Deploy_App job which are no longer required.

Also adds the delpoy_efg.yaml to the  deploy_app manifest.

It also triggers a build of 'efg-training' from the Deploy_App job
because Jenkins will not trigger a job if the choice is invalid. Also sets up
the correct parameters for the trigger, and reorders the structure of
the file.