Upgrade the ruby version used by puppet to 2.1.8
We're currently running a very old, unsupported, ruby version and
we should upgrade to a newer, still security patch supported one.

2.1.8 was picked as it's quite new (comparatively speaking) and we have
packages built for it.

Install ruby 2.1.8 everywhere as we're going to use it for puppet

We're leaving 1.9.3 in for now while we upgrade the puppet used ruby.
We're also unsure if we have any other scripts running that use this version.

Bump the puppet version