Start enabling future parser
The future parser will help us upgrade to Puppet 4. For now enable
it on the Puppetmaster. There are no changes when running Puppet
locally on the Puppetmaster with the future parser on.

Similarly, the jumpbox is a very simple machine.

The `stringify_facts` option is required because by default all facter
facts will be returned as strings which is no good when the future
parser is checking types.

The comment about booleans being weird doesn't add anything.