Update Jenkins deploy app tasks for reduced ambiguity
Currently users of Jenkins are often confused by the task names in the
deploy namespace and expect them to perform a deploy as well as the
associated restart or migrate task.

This change updates the commands in favour of renamed ones that are
easier to understand with updated description to aid people in picking
their task.