Provide a list of mongodb hosts for manuals-publisher
Ruby 2.1.* cannot parse the comma delimited list of mongodb uris
and we would like to be able to eliminate recent ruby-version
upgrades from problems we're seeing on Manuals Publisher by
reverting to ruby 2.1.2 as this is the version the codebase is
known to run happily on.
So provide a list of mongodb host:port strings that can be parsed
from a ENV var into yaml.