Schedule Auto PostgreSQL Backups
What
Auto postgreSQL backups run daily (every morning) via cron and fail.
The error output:
`pg_dump: [archiver (db)] query failed: ERROR:  cache lookup failed for index 506280`

It's not entirely clear but the backup runs while the envsync is in progress, which would cause
some indexes to not be available. This is further clarified by running the script exiting with no
errors after the envsync has completed.

How
Move the main backup script from cron daily to `/usr/local/bin`.
Create a cron job to run the script at `08:30`, `07:30` UTC time.
The alert we're trying to rectify only occurs in Staging and Integration, so we set the backup to run at `6:30`, `5:30` UTC
in production which is the time it would normally run with `cron.daily`.