Replace envvars with a source file
We set environmental variables to be used by the script but they weren't
actually required as they were already being set as variables and
passed into cURL as JSON.

This removes the environmental variables and creates a source file. This
means that they're not hardcoded in the script for anyone to read, and
can easily be pulled in to be set in the JSON data. Permissions on the
source file should be restrictive.