Add PostgreSQL WAL-E restore script
While writing documentation it occurred to me that we did not have a
basic restore script for restores from the same environment. This commit
adds the script.

It also removes the env_sync details added to Integration that was used
for testing restores from Production. We are currently not using this
method so it should be removed until we decide to use it.