Enable future parser
This enables to help test the upgrade to Puppet 4.

The `future_parser` config is only required on the puppetmaster, but it
doesn't hurt to have it on the client config.

The `stringify_facts` config fixes any problems we have with how data types are
evaluated in Puppet 4.

The `strict_variables` config is set to help us debug issues easier
because Puppet 4 has a stricter checking for undefined variables.

We're just setting the option for now rather than turning it on
anywhere.