Explicitly specify service URLs for auth-proxy.
Plek now automatically inserts a "draft-" prefix for all service URLs
when called on the draft-cache machines because they have the
PLEK_HOSTNAME_PREFIX env var set. It's not possible to override that
with an empty value, so instead set the two service domains that
authenticating-proxy actually cares about - errbit and signon -
explicitly.