Add Machine, Processes and Bouncer dashboards
I built these on the production grafana instance, but adding them to puppet for
persistence, and so that they appear in all environments.

The Machine and Processes dashboards are generic dashboards to provide easier
access to the metrics in Graphite about machine and app process metrics
respectively.

I created the Bouncer dashboard while investigating a high number of errors in
production.