Disable "support app background processing" check in integration
This check requires people to send test messages through the support queues to
actually make it pass, which they don't, which leads to be being in an UNKNOWN
state for 60 days at a time. Disabling this monitoring does risk hiding an actual failiure
but it's currentlly broken 99% of the time anyway so people just ack it and move on.

It's also currently disabled in staging.