Move S3 attachments sync to evening
The S3 attachment sync causes the machine to run extremely low on
memory, which *may* have caused problems with the asset-master to
connect to it. The asset-master virus scanning process copies each file
it scans to each slave, and if it's unable to connect it will eventually
time out, but it will also cause files to backup in processing if there
are a lot of them.

This makes it less likely that we will see an impact of this during the
daytime when there are more files being to uploaded for processing.